// src/types/rarity.model.ts
export interface RawRarity {
  UEcharcelllarge: number;
  UEcharframe: number;
  UEcharframesmall: number;
  UEdownback: number;
  UEframeid: number;
  UEframeidsmall: number;
  breakLevelBackPolygon: number;
  breakLevelBackTriangle: number;
  charcelllarge: number;
  charframe: number;
  charframesmall: number;
  downback: number;
  drawinfo: number;
  evolutionEffect: number;
  frameid: number;
  frameidsmall: number;
  id: number;
  imgbigid: number;
  imgid: number;
  imgroleid: number;
  rankllarge: number;
  triangle: number;
  trianglelarge: number;
  yardCell: number;
}

export interface Rarity {
  id: number;
  name: string | null;
  ueCharCellLarge: string | null;
  ueCharFrame: string | null;
  ueCharFrameSmall: string | null;
  ueDownBack: string | null;
  ueFrameId: string | null;
  ueFrameIdSmall: string | null;
  breakLevelBackPolygon: string | null;
  breakLevelBackTriangle: string | null;
  charCellLarge: string | null;
  charFrame: string | null;
  charFrameSmall: string | null;
  downBack: string | null;
  drawInfo: string | null;
  evolutionEffect: string | null;
  frameId: string | null;
  frameIdSmall: string | null;
  imgBigId: string | null;
  imgId: string | null;
  imgRoleId: string | null;
  ranklLarge: string | null;
  triangle: string | null;
  triangleLarge: string | null;
  yardCell: string | null;
}